LayoutTests/platform/ios-sim-deprecated/fast/events/pageshow-pagehide-expected.txt

frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
frame "<!--framePath //<!--frame0-->/<!--frame0-->-->" - has 1 onunload handler(s)
frame "<!--framePath //<!--frame0-->/<!--frame0-->-->" - has 1 onunload handler(s)
frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
Test pageshow/pagehide event behavior in subframes.

 ***Original load - onload and pageshow events should fire for subframes, and then for main frame***
Subsubframe window.onload
Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = false
Subframe window.onload
Subframe window.onpageshow, target = [object HTMLDocument], persisted = false
Main frame window.onload
Main frame window.onpageshow, target = [object HTMLDocument], persisted = false
***Navigating bottom-level subframe, onpagehide events should fire for subsubframe***
Subsubframe window.onpagehide, target = [object HTMLDocument], persisted = false
Subsubframe window.onunload
Subsubframe window.onload
Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = false
***Navigating mid-level subframe, onpagehide events should fire for both subframes***
Subframe window.onpagehide, target = [object HTMLDocument], persisted = false
Subframe window.onunload
Subsubframe window.onpagehide, target = [object HTMLDocument], persisted = false
Subsubframe window.onunload
Subframe window.onload
Subframe window.onpageshow, target = [object HTMLDocument], persisted = false
***Done***
